Beginning Required Role skills
So this is the first commit in the line to solve the problem of "how do we
actually know this user can do that".

At heatsync, we have a lot of tools that simply shouldn't be used unless you
are well versed in them (google "lathe accidents", even though it's NSFL).
These will be controlled soon by RFID interlocks with a central DB to manage
them. HSM should be that DB, but righ tnow anyone can just say "I AM TEH
AWESOME AT LATHE-- *dead*" and that would suck.

I'm going to be setting this up so that only certain users can grant certain
skills to other users. This will replace the @user.can_fly magick that
Trever added when we were still trying to figure out how permissions would
work, at least for this usecase. .can_fly magick can still be used for
administravia on the site itself, I think, since "adding users to HSM" skill
isn't really optimal.